[feature] Option for resuming and with/without sparse input & [bug fixed] Load checkpoint in cpu to avoid cuda out of memory

+++ train.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
-TEST=true
+TEST=false
 if [ $TEST = true ] ; then
     GPUS='7'
-    ARGS="--model_name test_parallel --num_workers 1 --load_weights_folder ~/tmp/mdp/models/weights_19"
+    ARGS="--model_name test_parallel --num_workers 1"
+    PORT=2334
 else
     GPUS='0,1,2,3,4,5,6'
-    ARGS="--num_workers 120"
+    ARGS="--num_workers 120 --load_weights_folder $HOME/tmp/mdp/models/weights_5 --start_epoch 5 "
+    PORT=2333
 fi
 
 NUM_GPUS=$(echo $GPUS | tr "," "\n" | wc -l)
@@ -19,7 +21,7 @@ export OMP_NUM_THREADS=$(($NUM_THREADS/$NUM_GPUS))
 python -m torch.distributed.launch \
  --nproc_per_node=$NUM_GPUS \
  --node_rank=0 \
- --master_port=2333 \
+ --master_port=$PORT \
  train.py \
  --gpus $GPUS \
  --png $ARGS
@@ -30,4 +32,6 @@ python -m torch.distributed.launch \
 #  --load_weights_folder models/mono_640x192 \
 #  --load_weights_folder ~/tmp/mdp/models/weights_19 \
 
-kill -9 $(ps aux | grep train.py | grep -v grep | awk '{print $2}')
+if [ $TEST = false ] ; then
+    kill -9 $(ps aux | grep train.py | grep -v grep | awk '{print $2}')
+fi
